Add usps in person proofing directory
