+++ app/forms/api/profile_creation_form.rb
@@ -1,5 +1,5 @@
 module Api
-  # todo: move this somewhere else, maybe the UspsInPersonProofer
+  # todo: move this somewhere else, maybe the UspsInPersonProofing::Proofer
   Applicant = Struct.new(
     :unique_id, :first_name, :last_name, :address, :city, :state, :zip_code,
     :email, keyword_init: true
